transfer MainActivity to HomeActivity
