+++ android_app/app/src/main/res/layout/activity_main.xml
@@ -6,18 +6,6 @@
     android:layout_height="match_parent"
     android:background="@color/cardview_light_background"
     tools:context=".MainActivity">
-    <FrameLayout
-        android:id="@+id/fragment_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_above="@+id/bottom_navigation"></FrameLayout>
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:background="?android:attr/windowBackground"
-        android:id="@+id/bottom_navigation"
-        app:menu="@menu/bottom_navigation">
-    </com.google.android.material.bottomnavigation.BottomNavigationView>
+
 
 </RelativeLayout>
